rclone backend: retry errors on load and store 3 times
Rclone internally retries requests, but not ones which stream data,
like load and store.

Since we have the data for load and store buffered in memory it is
easy to retry them.

The rclone server returns 500 errors for backend errors or internet
errors or protocol errors which might be fixed with a retry.