remove MStore from master branch for now
We first have to develop the Store better before building
stuff on top of it. MStore was rather in the way of doing
that.

Before ripping MStore out, I made a branch "mstore" with
the current state (as in master branch before this commit),
so it can be developed there.