use rclone:remote:path as "URL"
remote:path is a string as defined by the rclone documentation.

these strings can be easy, but they also can get rather complex, see:
https://rclone.org/docs/#syntax-of-remote-paths

if they get complex, the whole thing won't look like a standard URL
pattern proto://servername/path anyway. also "remote" is not a servername
(like something that can be resolved in DNS) but just a name of an entry
in the rclone configuration.